add diagnostic endpoint
